feat(gh-auth): serve the token endpoint from a Pages Function, not a Worker
The repo's Cloudflare token covers Pages (the site deploys with it) but not
Workers Scripts: 'wrangler deploy' fails with Authentication error [code: 10000]
'Failed to automatically retrieve account IDs', i.e. the missing authorization is
a Workers-scoped token, not a bug. Pages Functions run on the same runtime, so the
endpoint ships with the site the token can already deploy, and the secret is set as
Pages project env by deploy.yml.

  site/functions/gh-auth.js       the handler (onRequest), same security model
  site/functions/gh-auth.test.mjs drives it against the real API (10 checks)
  gh-auth-verify.yml              runs that harness on every PR to this path
  deploy.yml                      sets APP_ID / APP_PRIVATE_KEY / AUTH_TOKEN as
                                  Pages secrets from the existing repo secrets

The standalone Worker variant (with its deploy workflow) stays in this branch's
history, ready for the day a Workers-scoped token exists.
