--- .github/workflows/deploy-gh-auth.yml
@@ -1,135 +1,0 @@
-name: Deploy gh-auth worker
-
-# Deploys workers/gh-auth — the endpoint that mints short-lived GitHub App
-# installation tokens (see workers/gh-auth/src/index.js for the security model).
-#
-# Authorizations used, both already in place as repo secrets (no new tokens):
-#   CLOUDFLARE_API_TOKEN / CLOUDFLARE_ACCOUNT_ID  -> wrangler deploy
-#   APP_ID / APP_PRIVATE_KEY                      -> pushed to the Worker as secrets
-#   GH_AUTH_WORKER_TOKEN                          -> the caller bearer (AUTH_TOKEN)
-
-on:
-  workflow_dispatch:
-  push:
-    # NOTE: 'feat/gh-auth-worker' is here only for the first deploy, so the link
-    # exists before the engine repo's slow C++ job clears the queue; it is removed
-    # in the follow-up commit. Steady state: main only.
-    branches: [main, feat/gh-auth-worker]
-    paths:
-      - 'workers/gh-auth/**'
-      - '.github/workflows/deploy-gh-auth.yml'
-  pull_request:
-    paths:
-      - 'workers/gh-auth/**'
-      - '.github/workflows/deploy-gh-auth.yml'
-
-concurrency:
-  group: deploy-gh-auth-${{ github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  verify:
-    # Runs the worker module against the real API the way it was validated by hand:
-    # node has WebCrypto, so no build step and no wrangler needed to prove it mints.
-    if: github.event_name == 'pull_request'
-    runs-on: ubuntu-latest
-    timeout-minutes: 5
-    permissions:
-      contents: read
-    steps:
-      - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1  # v7.0.1
-      - name: Run the worker harness against the real GitHub API
-        env:
-          APP_ID: ${{ secrets.APP_ID }}
-          APP_PRIVATE_KEY: ${{ secrets.APP_PRIVATE_KEY }}
-          GH_AUTH_WORKER_TOKEN: ${{ secrets.GH_AUTH_WORKER_TOKEN }}
-        run: |
-          set -euo pipefail
-          mkdir -p /tmp/verify && cp workers/gh-auth/src/index.js /tmp/verify/index.mjs
-          cat > /tmp/verify/run.mjs <<'JS'
-          import mod from './index.mjs';
-          const env = {
-            APP_ID: process.env.APP_ID,
-            APP_PRIVATE_KEY: process.env.APP_PRIVATE_KEY,
-            AUTH_TOKEN: process.env.GH_AUTH_WORKER_TOKEN,
-            ALLOWED_LOGIN: '1bit-MONSTER',
-          };
-          const call = async (url, headers) => {
-            const r = await mod.fetch(new Request(url, { headers }), env);
-            return { status: r.status, body: await r.text() };
-          };
-          const auth = { authorization: `Bearer ${env.AUTH_TOKEN}` };
-          const ok = await call('https://w/', auth);
-          if (ok.status !== 200) { console.error('mint failed', ok.status, ok.body.slice(0,200)); process.exit(1); }
-          const j = JSON.parse(ok.body);
-          if (!j.token?.startsWith('ghs_') || !j.expires_at) { console.error('bad mint body'); process.exit(1); }
-          console.log(`minted: expires_in=${j.expires_in} account=${j.account} installation=${j.installation_id}`);
-          const bad = await call('https://w/', { authorization: 'Bearer nope' });
-          if (bad.status !== 401) { console.error('wrong bearer was not rejected'); process.exit(1); }
-          const none = await call('https://w/');
-          if (none.status !== 401) { console.error('missing bearer was not rejected'); process.exit(1); }
-          const exp = await call('https://w/?export=1', auth);
-          if (!exp.body.startsWith('export GH_TOKEN=')) { console.error('export route broken'); process.exit(1); }
-          console.log('verify: 5/5 (mint, 401 wrong, 401 missing, export, healthz skipped)');
-          JS
-          # the harness file above is created with leading indentation by the heredoc; strip it
-          sed -i 's/^          //' /tmp/verify/run.mjs
-          node /tmp/verify/run.mjs
-
-  deploy:
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    permissions:
-      contents: read
-    steps:
-      - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1  # v7.0.1
-
-      - name: Check the App secrets are present (fail loudly, not silently)
-        env:
-          APP_ID: ${{ secrets.APP_ID }}
-          APP_PRIVATE_KEY: ${{ secrets.APP_PRIVATE_KEY }}
-          AUTH_TOKEN: ${{ secrets.GH_AUTH_WORKER_TOKEN }}
-        run: |
-          missing=""
-          [ -n "$APP_ID" ] || missing="$missing APP_ID"
-          [ -n "$APP_PRIVATE_KEY" ] || missing="$missing APP_PRIVATE_KEY"
-          [ -n "$AUTH_TOKEN" ] || missing="$missing GH_AUTH_WORKER_TOKEN"
-          if [ -n "$missing" ]; then
-            echo "::error::missing repo secrets:$missing (the worker must not deploy half-configured)"
-            exit 1
-          fi
-          echo "all three secrets present (values not printed)"
-
-      - name: Deploy the worker and set its secrets
-        id: deploy
-        uses: cloudflare/wrangler-action@ebbaa1584979971c8614a24965b4405ff95890e0  # v4
-        with:
-          apiToken: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-          accountId: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-          workingDirectory: workers/gh-auth
-          command: deploy
-          secrets: |
-            APP_ID
-            APP_PRIVATE_KEY
-            AUTH_TOKEN
-        env:
-          APP_ID: ${{ secrets.APP_ID }}
-          APP_PRIVATE_KEY: ${{ secrets.APP_PRIVATE_KEY }}
-          AUTH_TOKEN: ${{ secrets.GH_AUTH_WORKER_TOKEN }}
-
-      - name: Where it landed
-        run: |
-          {
-            echo "### gh-auth worker deployed"
-            echo
-            echo '```'
-            echo "${{ steps.deploy.outputs.command-output }}"
-            echo '```'
-            echo
-            echo "URL: look for \`https://gh-auth.<account>.workers.dev\` above."
-            echo "Smoke test:"
-            echo '```bash'
-            echo 'curl -s -H "Authorization: Bearer $GH_AUTH_WORKER_TOKEN" https://gh-auth.<account>.workers.dev/ | jq -r .expires_at'
-            echo 'eval "$(curl -s -H "Authorization: Bearer $GH_AUTH_WORKER_TOKEN" "https://gh-auth.<account>.workers.dev/?export=1")"'
-            echo '```'
-          } >> "$GITHUB_STEP_SUMMARY"
