--- workers/gh-auth/wrangler.toml
@@ -1,14 +1,0 @@
-name = "gh-auth"
-main = "src/index.js"
-compatibility_date = "2026-04-20"
-
-# Secrets (never in this file — set by .github/workflows/deploy-gh-auth.yml from
-# the repo's Actions secrets):
-#   APP_ID            GitHub App id (reuses the repo's existing APP_ID secret)
-#   APP_PRIVATE_KEY   App private key, PKCS#8 PEM (reuses APP_PRIVATE_KEY)
-#   AUTH_TOKEN        bearer that callers must present (repo secret GH_AUTH_WORKER_TOKEN)
-
-[vars]
-# Which App installation to mint for. The App is installed on 1bit-MONSTER and
-# sees 5 repos; if the installation id is known it can be pinned with APP_INSTALL_ID.
-ALLOWED_LOGIN = "1bit-MONSTER"
